update cron job api. Fundamentally a CronJob needs the following pieces:
A schedule (the cron in CronJob)
A template for the Job to run (the job in CronJob)
We’ll also want a few extras, which will make our users’ lives easier:

A deadline for starting jobs (if we miss this deadline, we’ll just wait till the next scheduled time)
What to do if multiple jobs would run at once (do we wait? stop the old one? run both?)
A way to pause the running of a CronJob, in case something’s wrong with it
Limits on old job history